Move GamePauseScreen into its own file
