+++ UI/MainScreen.h
@@ -76,42 +76,6 @@ private:
 	bool UseVerticalLayout() const;
 };
 
-class GamePauseScreen : public UIDialogScreenWithGameBackground {
-public:
-	GamePauseScreen(const std::string &filename) : UIDialogScreenWithGameBackground(filename), saveSlots_(NULL) {}
-	virtual ~GamePauseScreen();
-
-	virtual void onFinish(DialogResult result);
-
-protected:
-	virtual void CreateViews();
-	virtual void update(InputState &input);
-	virtual void sendMessage(const char *message, const char *value);
-	void CallbackDeleteConfig(bool yes);
-
-private:
-	UI::EventReturn OnMainSettings(UI::EventParams &e);
-	UI::EventReturn OnGameSettings(UI::EventParams &e);
-	UI::EventReturn OnExitToMenu(UI::EventParams &e);
-	UI::EventReturn OnReportFeedback(UI::EventParams &e);
-
-	UI::EventReturn OnSaveState(UI::EventParams &e);
-	UI::EventReturn OnLoadState(UI::EventParams &e);
-	UI::EventReturn OnRewind(UI::EventParams &e);
-
-	UI::EventReturn OnStateSelected(UI::EventParams &e);
-	UI::EventReturn OnCwCheat(UI::EventParams &e);
-
-	UI::EventReturn OnCreateConfig(UI::EventParams &e);
-	UI::EventReturn OnDeleteConfig(UI::EventParams &e);
-
-	UI::EventReturn OnSwitchUMD(UI::EventParams &e);
-
-	UI::ChoiceStrip *saveSlots_;
-	UI::Choice *saveStateButton_;
-	UI::Choice *loadStateButton_;
-};
-
 class UmdReplaceScreen : public UIDialogScreenWithBackground {
 public:
 	UmdReplaceScreen() {}
